# docs/openapi/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Antaride",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api"
        }
    ],
    "paths": {
...
        "/v1/auth/otp/request": {
            "post": {
                "operationId": "v1.auth.otp.request",
                "description": "Response-nya SELALU sama bentuknya, terdaftar maupun belum.",
                "summary": "Minta kode OTP",
                "tags": [
                    "Otp"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestOtpRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "data": {
                                            "anyOf": [
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "phone_masked": {
                                                            "type": "string",
                                                            "description": "Nomor dikembalikan dalam bentuk tersamarkan, bukan penuh.\nAplikasi sudah tahu nomor yang dia kirim; yang ditampilkan di\nlayar \"masukkan kode\" adalah konfirmasi, dan nomor penuh di sana\nhanya menambah data yang bisa dibaca orang di sekitar."
                                                        },
                                                        "purpose": {
                                                            "type": "string"
                                                        },
                                                        "expires_at": {
                                                            "type": "string"
                                                        },
                                                        "expires_in_seconds": {
                                                            "type": [
                                                                "object",
                                                                "null"
                                                            ]
                                                        },
                                                        "resend_after_seconds": {
                                                            "type": "integer"
                                                        },
                                                        "debug_code": {
                                                            "type": [
                                                                "string",
                                                                "null"
                                                            ]
                                                        }
                                                    },
                                                    "required": [
                                                        "phone_masked",
                                                        "purpose",
                                                        "expires_at",
                                                        "expires_in_seconds",
                                                        "resend_after_seconds",
                                                        "debug_code"
                                                    ]
                                                },
                                                {
                                                    "type": "object",
                                                    "additionalProperties": {
                                                        "type": "string"
                                                    }
                                                },
                                                {
                                                    "type": "string"
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RequestOtpRequest": {
                "type": "object",
                "description": "Minta kode OTP.",
                "properties": {
                    "phone": {
                        "type": "string",
                        "description": "Validasi bentuknya SENGAJA longgar di sini. Normalisasi dan pemeriksaan bentuk yang sebenarnya ada di\n`PhoneNumber::normalize()`, dan itu satu-satunya tempat yang tahu\nbentuk kanoniknya. Menuliskan regex nomor HP di sini berarti ada\nDUA definisi \"nomor HP yang valid\" yang harus sepakat \u2014 dan yang\ndi FormRequest akan tertinggal, sehingga nomor yang sah ditolak\nsebelum Action-nya sempat menormalkannya.\n\nYang diperiksa di sini hanya yang tidak mungkin: terlalu pendek\nuntuk nomor apa pun, atau terlalu panjang untuk masuk akal.",
                        "minLength": 9,
                        "maxLength": 20
                    },
                    "purpose": {
                        "type": "string",
                        "enum": [
                            "login",
                            "register",
                            "change_phone"
                        ]
                    }
                },
                "required": [
                    "phone"
                ],
                "title": "RequestOtpRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}